[AN] refactor(#139): 패키지 이름 변경
login -> auth
