+++ android/gitudy/presentation/src/main/java/com/takseha/presentation/ui/login/SocialLoginCompleteActivity.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
-import com.takseha.data.dto.login.login.Role
+import com.takseha.data.dto.auth.login.Role
 import com.takseha.presentation.R
 import com.takseha.presentation.databinding.ActivitySocialLoginCompleteBinding
 import com.takseha.presentation.ui.home.MainHomeActivity
